NRFFOSDK-6505: Removed init_packet.py.
